+++ lib/core/apis/EcoleDirecte/ecoleDirecteCloud.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:http/http.dart' as http;
 import 'package:ynotes/core/apis/EcoleDirecte.dart';
 import 'package:ynotes/core/logic/modelsExporter.dart';
-import 'package:ynotes/ui/screens/cloud/cloudPage.dart';
+import 'package:ynotes/ui/screens/cloud/index.dart';
 
 //The basical function to change folder
 Future<List<CloudItem>?> changeFolder(String path) async {
@@ -16,8 +16,7 @@ Future<List<CloudItem>?> changeFolder(String path) async {
   finalPath.forEach((item) {
     concatenate.write(r'\' + item);
   });
-  var url =
-      'https://api.ecoledirecte.com/v3/cloud/W/$cloudUsedFolder.awp?verbe=get&idFolder=$concatenate';
+  var url = 'https://api.ecoledirecte.com/v3/cloud/W/$cloudUsedFolder.awp?verbe=get&idFolder=$concatenate';
   url = Uri.encodeFull(url);
   print(url);
   List<CloudItem> toReturn = [];
@@ -26,9 +25,7 @@ Future<List<CloudItem>?> changeFolder(String path) async {
   String data = 'data={"token": "$token"}';
 
   var body = data;
-  var response = await http
-      .post(Uri.parse(url), headers: headers, body: body)
-      .catchError((e) {
+  var response = await http.post(Uri.parse(url), headers: headers, body: body).catchError((e) {
     throw ("Impossible de se connecter. Essayez de vérifier votre connexion à Internet ou réessayez plus tard.");
   });
 
@@ -43,9 +40,7 @@ Future<List<CloudItem>?> changeFolder(String path) async {
               element["libelle"],
               element["type"].toString().toUpperCase(),
               element["proprietaire"] != null
-                  ? element["proprietaire"]["prenom"] +
-                      " " +
-                      element["proprietaire"]["nom"]
+                  ? element["proprietaire"]["prenom"] + " " + element["proprietaire"]["nom"]
                   : "",
               false,
               element["date"],
