+++ lib/main.dart
@@ -16,8 +16,8 @@ import 'package:ynotes/core/services/notifications.dart';
 import 'package:ynotes/globals.dart';
 import 'package:ynotes/router.dart';
 import 'package:ynotes/ui/components/hive_life_cycle_manager.dart';
-import 'package:ynotes/ui/screens/carousel/carousel.dart';
-import 'package:ynotes/ui/screens/loading/loadingPage.dart';
+import 'package:ynotes/ui/screens/carousel/index.dart';
+import 'package:ynotes/ui/screens/loading/index.dart';
 import 'package:ynotes/ui/screens/summary/summaryPage.dart';
 
 import 'core/utils/themeUtils.dart';
