+++ lib/router.dart
@@ -2,12 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:ynotes/core/logic/appConfig/models.dart';
 import 'package:ynotes/ui/screens/agenda/index.dart';
-import 'package:ynotes/ui/screens/carousel/carousel.dart';
-import 'package:ynotes/ui/screens/cloud/cloudPage.dart';
-import 'package:ynotes/ui/screens/downloads/downloadsPage.dart';
+import 'package:ynotes/ui/screens/carousel/index.dart';
+import 'package:ynotes/ui/screens/cloud/index.dart';
+import 'package:ynotes/ui/screens/downloads/index.dart';
 import 'package:ynotes/ui/screens/error_page.dart';
-import 'package:ynotes/ui/screens/grades/gradesPage.dart';
-import 'package:ynotes/ui/screens/homework/homeworkPage.dart';
+import 'package:ynotes/ui/screens/grades/index.dart';
+import 'package:ynotes/ui/screens/homework/index.dart';
 import 'package:ynotes/main.dart';
 import 'package:ynotes/ui/screens/mail/mailPage.dart';
 import 'package:ynotes/ui/screens/polls/pollsPage.dart';
