+++ lib/ui/components/dialogs.dart
@@ -11,15 +11,15 @@ import 'package:ynotes/core/utils/themeUtils.dart';
 import 'package:ynotes/ui/components/dialogs/authorizations_dialog.dart';
 import 'package:ynotes/ui/components/dialogs/color_picker.dart';
 import 'package:ynotes/ui/components/dialogs/update_note_dialog.dart';
-import 'package:ynotes/ui/screens/agenda/widgets/persistantNotificationDialog.dart';
-import 'package:ynotes/ui/screens/grades/gradesPageWidgets/shareGradeDialog.dart';
-import 'package:ynotes/ui/screens/homework/homeworkPageWidgets/homeworkDetails.dart';
-import 'package:ynotes/ui/screens/login/loginPageWidgets/pronoteLocationDialog.dart';
+import 'package:ynotes/ui/screens/agenda/widgets/persistant_notification_dialog.dart';
+import 'package:ynotes/ui/screens/grades/widgets/share_grade_dialog.dart';
+import 'package:ynotes/ui/screens/homework/widgets/homework_details.dart';
+import 'package:ynotes/ui/screens/login/widgets/pronote_location_dialog.dart';
 import 'package:ynotes/ui/screens/mail/mailPageWidgets/writeMailBottomSheet.dart';
 
 import '../../useful_methods.dart';
-import '../screens/agenda/widgets/recurringEventsDialog.dart';
-import '../screens/downloads/downloadsPageWidgets/folderChoiceDialog.dart';
+import '../screens/agenda/widgets/recurring_events_dialog.dart';
+import '../screens/downloads/widgets/folder_choice_dialog.dart';
 import '../screens/mail/mailPageWidgets/newRecipientDialog.dart';
 import 'dialogs/multiple_choices_dialog.dart';
 import 'dialogs/number_choice_dialog.dart';
