+++ lib/ui/screens/carousel/index.dart
